+++ h7/can/can.h
@@ -24,7 +24,7 @@ enum class Peripheral {
 };
 
 
-constexpr int peripheral_count = 2;
+constexpr size_t peripheral_count = 2;
 
 
 struct RxPinConfig {
